[ui] Create component embed system
[ui] Create component embed system
- Component embed are components that are created by the component
implementation itself.  It is used to achieve code reuse.
- Create new arguments and modify existing arguments.
- Update XSD.
- Update code and XML usages.

[ui] Complete 'UIWindowComponent'
- Use the component embed system in 'UIWindowComponent'.

[ui] Create 'UIShapeComponent'
- It draws a shape.  That is it.

Miscellaneous
- Improve code.
- Refactor code.

+++ subprojects/ui/src/main/java/io/github/etaoinshrdlcumwfgypbvkjxqz/capablecables/ui/impl/animations/controls/UIStandardAnimationControlFactory.java
@@ -7,7 +7,7 @@ import io.github.etaoinshrdlcumwfgypbvkjxqz.capablecables.utilities.systems.time
 public enum UIStandardAnimationControlFactory {
 	;
 
-	public static final int INFINITE_LOOP = -1;
+	public static final long INFINITE_LOOP = -1L;
 
 	public static UIAbstractStandardAnimationControl createSimple(EnumDirection direction,
 	                                                              IUIAnimationTarget target,
@@ -85,7 +85,7 @@ public enum UIStandardAnimationControlFactory {
 		}
 	}
 
-	public static int getInfiniteLoop() { return INFINITE_LOOP; }
+	public static long getInfiniteLoop() { return INFINITE_LOOP; }
 
 	public enum EnumDirection {
 		NORMAL,
